Move the crytpo headers out of public/internal and into private folders within src.
